+++ src/components/generic/editable-field-components/index.tsx
@@ -1,5 +1,5 @@
 export { default as MetaformTextFieldComponent } from "./metaform-text-field-component";
 export { default as MetaformHtmlComponent } from "./html-field-component";
-export { default as MetaformRadioFieldComponent } from "./metaform-radio-field-component";
+export { default as MetaformRadioFieldComponent } from "./radio-field-component";
 export { default as MetaformSubmitFieldComponent } from "./metaform-submit-field-component";
 export { default as MetaformNumberFieldComponent } from "./number-field-component";
